+++ packages/scripts/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.3.0 (Unreleased)
+
+### Improvements
+
+- New flag `--ignore` for `check-licenses` script
+
 ## 2.2.1 (2018-09-05)
 
 ### Bug Fix
@@ -7,7 +13,6 @@
 ### Polish
 
 - Updated dependencies: `jest`, `npm-package-json-lint` and `read-pkg-up`
-
 
 ## 2.0.0 (2018-07-12)